Create servicios en proceso

+++ backend/controllers/ImgprofileController.php
@@ -98,7 +98,9 @@ class ImgprofileController extends Controller
     public function actionDelete($id)
     {
         $imgBorrar = $this->findModel($id)->url;
-        unlink($imgBorrar);
+        if (!strstr($imgBorrar, 'imgAva')) {
+            unlink($imgBorrar);
+        }
         $this->findModel($id)->delete();
     }
